int is better than short.
git-svn-id: http://svn.code.sf.net/p/dnsjava/code/trunk@1411 c76caeb1-94fd-44dd-870f-0c9d92034fc1

+++ org/xbill/DNS/Name.java
@@ -576,7 +576,7 @@ byteString(byte [] array, int pos) {
 	StringBuffer sb = new StringBuffer();
 	int len = array[pos++];
 	for (int i = pos; i < pos + len; i++) {
-		short b = (short)(array[i] & 0xFF);
+		int b = array[i] & 0xFF;
 		if (b <= 0x20 || b >= 0x7f) {
 			sb.append('\\');
 			sb.append(byteFormat.format(b));
